renamed files to reflect decision to user simpler model for managing semantic convention versions
