+++ src/Shared/AWS/AWSSemanticConventions.Base.cs
