+++ src/Shared/AWS/AWSSemanticConventions.v1.10.cs
