+++ src/Shared/AWS/AWSSemanticConventions.v1.10_1.cs
